requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
pillow==11.0.0
python-telegram-bot==21.6
sniffio==1.3.1
